pyproject.toml
[project]
name = "python-vuln-demo"
version = "0.1.0"
description = "Intentionally vulnerable Python demo app managed with uv for security scanning demos."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "flask==0.12",
  "requests==2.19.1",
  "PyYAML==5.4",
  "django==2.0.0",
  "jinja2<3.1.0",
  "werkzeug<1.0.0",
  "itsdangerous<1.0.0"
]